задание_повышенной_сложности_1/python_client/requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
pytest>=7.4.3
